Combat fixes, added COMBATF_PREHIT_NORANGE - Fixed: COMBATF_PREHIT didn't fire the triggers in the correct order, also fixed a possible exploit while using PreHit (issue #108) - Added: COMBATF_PREHIT_NORANGE to allow starting the close-combat swing at distance; the blow will still be landed only when entering in range. - Fixed: After each combat hit the current combat skill was halted (actually it failed), then the same action was started again. This resulted in 	the @Fail, @SkillFail, @Start, @SkillStart triggers to be fired at each hit. 	- Fixed: NPCs doing extra AI actions (wandering, food, etc checks) after each hit in combat. This caused sometimes the current combat action to be resetted 	and then forced again, erasing the NPC combat data, which ultimately results in the NPC hitting twice in a few tenths of seconds.

+++ src/game/chars/CCharStat.cpp
@@ -371,7 +371,7 @@ bool CChar::Stats_Regen(int64 iTimeDiff)
 		{
 			int iGain = Skill_Focus(i);
 			if (iGain > 0)
-			     mod += iGain;
+			     mod += (short)iGain;
 		}
 		short StatLimit = Stat_GetMax(i);
